% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_distance.R
\name{.validate_lat_lon_vector}
\alias{.validate_lat_lon_vector}
\title{Validate vector lat/lon}
\usage{
.validate_lat_lon_vector(lat, lon)
}
\arguments{
\item{lat}{[numeric]: Latitudes in decimal degrees.}

\item{lon}{[numeric]: Longitudes in decimal degrees.}
}
\description{
Validate vector lat/lon
}
\keyword{internal}
