% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git}
\alias{git}
\title{Running a Git Command}
\usage{
git(
  ...,
  input = NULL,
  working_dir = NULL,
  throw_on_stderr = FALSE,
  env = NULL,
  just_the_proc = FALSE
)
}
\arguments{
\item{...}{the arguments pass to git command, supports the
\href{https://rlang.r-lib.org/reference/nse-force.html}{big bang} operator \verb{!!!}}

\item{input}{text pass to stdin}

\item{working_dir}{working directory}

\item{throw_on_stderr}{throw an error if git command executed successfully but
stderr is not empty}

\item{env}{additional environment variable pass to git}

\item{just_the_proc}{return the underlying process object instead of the output}
}
\value{
The output in a scalar character. It may contain a trailing newline. Use \code{as.character()}
to ensure the trailing newline is trimmed.
}
\description{
Running a Git Command
}
\details{
If \code{git()} fails to locate the git binary automatically, the path could be
specify with \code{options(git_path = "/path/to/git")}.
}
\examples{
\donttest{
td <- file.path(tempdir(), "git-examples")
dir.create(td, showWarnings = FALSE)
setwd(td)

git("init")
git("config", "--local", "user.email", "you@example.com")
git("config", "--local", "user.name", "Your Name")


cat("hello\n", file = "hello.txt")

git("status")

git("add", "hello.txt")

git("commit", "-m", "initial message")

git("log")

cat("world\n", file = "world.txt")

git("add", "world.txt")

git("commit", "-F", "-", input = "second message")

(sha <- git("rev-parse", "HEAD"))

git("show", sha)
}

}
