% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freida_program_loader.R
\name{inference_output_columns}
\alias{inference_output_columns}
\title{Expected columns in find_nearest_program() output}
\usage{
inference_output_columns()
}
\value{
Character vector of required column names
}
\description{
Expected columns in find_nearest_program() output
}
\keyword{internal}
