% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freida_program_loader.R
\name{match_street_address}
\alias{match_street_address}
\title{Match Street Addresses}
\usage{
match_street_address(phys_addr, prog_addr)
}
\arguments{
\item{phys_addr}{Normalized address list (from normalize_street_address)}

\item{prog_addr}{Normalized address list (from normalize_street_address)}
}
\value{
Logical
}
\description{
Compares physician address to program address. Returns TRUE if
street number matches AND at least one key street token matches.
}
\keyword{internal}
