% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfd_prevalence.R
\name{pfd_prevalence_by_age}
\alias{pfd_prevalence_by_age}
\title{Age-specific PFD prevalence (vectorized). Ages below 0 return 0.}
\usage{
pfd_prevalence_by_age(age)
}
\description{
Age-specific PFD prevalence (vectorized). Ages below 0 return 0.
}
