% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freida_program_loader.R
\name{resolve_freida_dir}
\alias{resolve_freida_dir}
\title{Resolve FREIDA data directory}
\usage{
resolve_freida_dir()
}
\value{
Character path to directory containing FREIDA CSVs
}
\description{
Checks three locations in priority order:
  1. FREIDA_DATA_DIR environment variable
  2. freida_data_dir key in config/paths.yml
  3. data/freida/ (project-local copy)
}
