% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_distance.R
\name{spatial_distance_module}
\alias{spatial_distance_module}
\title{Spatial Distance Utilities}
\description{
Canonical implementations of spatial distance calculations. Consolidates
scattered \code{haversine_distance()} implementations and adds safety guards.
}
\seealso{
Other geospatial-utils: 
\code{\link{haversine_km_vec}()}
}
\concept{geospatial-utils}
