% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workforce_cliff_engine.R
\name{wc_active_ages}
\alias{wc_active_ages}
\title{Primary active-age vectors per cohort (ret==FALSE + ABU net-new appended to URPS).}
\usage{
wc_active_ages(coh, abu_age = wc_load_abu_ages())
}
\arguments{
\item{coh}{[data.frame]: Cohort from `wc_load_cohort()`.}

\item{abu_age}{[integer]: ABU urology-pathway ages to append for the both-pathway baseline.}
}
\description{
Primary active-age vectors per cohort (ret==FALSE + ABU net-new appended to URPS).
}
