% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workforce_cliff_engine.R
\name{wc_load_abu_ages}
\alias{wc_load_abu_ages}
\title{ABU urology-pathway net-new active URPS ages (age from ABU certification year).}
\usage{
wc_load_abu_ages()
}
\description{
ABU urology-pathway net-new active URPS ages (age from ABU certification year).
}
