% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliff-package.R
\docType{package}
\name{cliff-package}
\alias{cliff}
\alias{cliff-package}
\title{cliff: Gynecologic Surgical Subspecialty Workforce Cliff Analysis}
\description{
Reproducible analysis and manuscript for near-term fellowship-completion versus clinical-practice-departure balance in gynecologic oncology and urogynecology (2025-2029). Extracted from the isochrones monorepo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mufflyt/cliff}
  \item Report bugs at \url{https://github.com/mufflyt/cliff/issues}
}

}
\author{
\strong{Maintainer}: Tyler Muffly \email{tyler.muffly@dhha.org} (\href{https://orcid.org/0000-0002-2044-1693}{ORCID})

}
\keyword{internal}
