% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git}
\alias{git}
\title{Running a Git Command}
\usage{
git(
  ...,
  input = NULL,
  working_dir = NULL,
  throw_on_stderr = FALSE,
  env = NULL,
  just_the_proc = FALSE
)
}
\arguments{
\item{...}{the arguments pass to git command}

\item{input}{text pass to stdin}

\item{working_dir}{working directory}

\item{throw_on_stderr}{throw an error if git command executed successfully but
stderr is not empty}

\item{env}{additional environment variable pass to git}

\item{just_the_proc}{return the underlying process object instead of the output}
}
\description{
Running a Git Command
}
