% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{miles_to_meters}
\alias{miles_to_meters}
\title{Convert statute miles to metres (vectorised; NA-preserving).}
\usage{
miles_to_meters(mi)
}
\arguments{
\item{mi}{[numeric]: Distance in miles.}
}
\description{
Convert statute miles to metres (vectorised; NA-preserving).
}
